Added things for mutt and a couple other tweaks

+++ install.sh
@@ -1,12 +1,23 @@
 #!/bin/sh
 
 if [ -z "$1" ] || [ -z "$2" ]; then
-  echo 'Usage: ./install.sh <full_name> <email>'
-  exit
+	echo 'Usage: ./install.sh <full_name> <email>'
+	exit
 else
-  GIT_NAME="$1"
-  GIT_EMAIL="$2"
+	FULL_NAME="$1"
+	EMAIL="$2"
 fi
+
+
+ask_if() (
+	printf 'Want %s? (y/n): ' "$1"
+	read CHOICE
+	if [ "$CHOICE" = 'y' ]; then
+		exit 0
+	else
+		exit 1
+	fi
+)
 
 
 #--------------------------------------------------------------------------------
@@ -21,23 +32,19 @@ mkdir -p ~/.vim/bundle
 curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
 echo " done"
 
-# vim-sensible
-echo -n 'installing vim-sensible...'
-rm -rf ~/.vim/bundle/vim-sensible
-git clone -q https://github.com/tpope/vim-sensible.git ~/.vim/bundle/vim-sensible
-echo " done"
+if ask_if 'javascript'; then
+	# vim-javascript
+	echo -n 'installing vim-javascript...'
+	rm -rf ~/.vim/bundle/vim-javascript
+	git clone -q https://github.com/pangloss/vim-javascript.git ~/.vim/bundle/vim-javascript
+	echo " done"
 
-# vim-javascript
-echo -n 'installing vim-javascript...'
-rm -rf ~/.vim/bundle/vim-javascript
-git clone -q https://github.com/pangloss/vim-javascript.git ~/.vim/bundle/vim-javascript
-echo " done"
-
-# vim-jsx
-echo -n 'installing vim-jsx...'
-rm -rf ~/.vim/bundle/vim-jsx
-git clone -q https://github.com/mxw/vim-jsx.git ~/.vim/bundle/vim-jsx
-echo " done"
+	# vim-jsx
+	echo -n 'installing vim-jsx...'
+	rm -rf ~/.vim/bundle/vim-jsx
+	git clone -q https://github.com/mxw/vim-jsx.git ~/.vim/bundle/vim-jsx
+	echo " done"
+fi
 
 cat << EOF > ~/.vimrc
 syntax on
@@ -66,6 +73,9 @@ autocmd FileType css            setlocal shiftwidth=4 tabstop=4 noexpandtab
 autocmd FileType sshconfig      setlocal shiftwidth=4 tabstop=4 noexpandtab
 autocmd FileType go             setlocal shiftwidth=4 tabstop=4 noexpandtab
 
+" markdown syntax is poorly defined so highlighting is a dumpster fire
+autocmd FileType markdown setlocal syntax=off
+
 execute pathogen#infect()
 filetype plugin indent on
 EOF
@@ -93,8 +103,8 @@ cat << EOF > ~/.gitconfig
 [color]
   ui = auto
 [user]
-  name = $GIT_NAME
-  email = $GIT_EMAIL
+  name = $FULL_NAME
+  email = $EMAIL
 EOF
 
 
@@ -111,8 +121,6 @@ EOF
 #--------------------------------------------------------------------------------
 # bash
 #--------------------------------------------------------------------------------
-
-
 
 cat << 'EOF' > ~/.bashrc
 umask 022
@@ -185,3 +193,70 @@ activate() {
 	. venv/bin/activate
 }
 EOF
+
+
+#--------------------------------------------------------------------------------
+# mutt
+#--------------------------------------------------------------------------------
+
+if ask_if 'email'; then
+	mkdir -p ~/.mutt/cache
+
+	if ! which mutt; then
+		printf 'warning: mutt is not installed\n'
+	fi
+
+	cat << EOF > ~/.muttrc
+set realname = "${FULL_NAME}"
+set from = "${EMAIL}"
+set use_from = yes
+
+set imap_user = "${EMAIL}"
+set folder = "imaps://imap.gmail.com/"
+set spoolfile = "+INBOX"
+set postponed = "+[Gmail]/Drafts"
+
+# gmail saves sent e-mail to +[Gmail]/Sent, so we don't want duplicates
+unset record
+set smtp_url = "smtps://$(printf ${EMAIL} | cut -d@ -f1)@smtp.gmail.com/"
+set ssl_force_tls = yes
+
+set header_cache =~/.mutt/cache/headers
+set message_cachedir =~/.mutt/cache/bodies
+set certificate_file =~/.mutt/certificates
+set move = no
+set imap_keepalive = 900
+
+# sort emails in index first by thread, then by date, newest first
+set sort = "threads"
+set sort_aux = "reverse-date"
+
+# don't mark unread messages that were previously seen but not read as 'O'
+unset mark_old
+
+# compose view options
+set edit_headers
+set forward_format = "Fwd: %s"
+set include
+set forward_quote
+set editor = "vim"
+# going to need to configure vim for this
+#set text_flowed
+# for when you want to turn mime messages into plain text when forwarding
+#unset mime_forward
+#set forward_decode
+
+# key bindings
+bind index j next-entry
+bind index k previous-entry
+bind index g first-entry
+bind index G last-entry
+bind index n search-next
+bind index N search-opposite
+
+bind pager j next-line
+bind pager k previous-line
+bind pager g top
+bind pager G bottom
+EOF
+fi
